tests/etl/requirements.txt
async-timeout==4.0.2
certifi==2022.12.7
elastic-transport==8.4.0
elasticsearch==7.9.1
psycopg2-binary==2.9.5
pydantic==1.10.5
python-dotenv==0.21.1
redis==4.5.1
typing_extensions==4.5.0
urllib3==1.26.14